refactor to 8 bits per limb + fixed bug in ecamples

+++ keccak-air/src/constants.rs
@@ -158,8 +158,8 @@ const RC_BITS: [[u8; 64]; 24] = [
     ],
 ];
 
-pub(crate) const fn rc_value_limb(round: usize, limb: usize) -> u16 {
-    (RC[round] >> (limb * BITS_PER_LIMB)) as u16
+pub(crate) const fn rc_value_limb(round: usize, limb: usize) -> u8 {
+    (RC[round] >> (limb * BITS_PER_LIMB)) as u8
 }
 
 pub(crate) const fn rc_value_bit(round: usize, bit_index: usize) -> u8 {
